Some other typos
I take the chance to ask... In the introduction, I felt like the first time I read the presentation at the bitDubai website. I read things like "wallet", "digital asset", "trust-less", "meta-data" being used as words that I should know already. What if I know absolutely nothing about those terms?

If I at least know that bitcoins is digital money that I can pass to someone else without using a bank and sort. What is the actual problem with wallets? Everyone can use Electrum or Coinbase, they won't get short of stock, or that's the thing I would understand, if you don't clear the 'wallet' term used here, or the exact problem with the actual wallets. The other words are more technical terms, so the average reader could have a (maybe wrong) idea or not at all. With all this I mean, if that's the introduction, shouldn't it have to _introduce_ the reader to what is all this about? Bitcoin is a lot to talk about, yes, but at least it could be explained the most simple way as possible so the reader knows what is the 'blockchain'. And the other words should be cleared too.