Refactor: Fix some exceptions occurred when the json got from swapbot site doesn't have any balance information.
Also, we included only for testing, right now, 18/04/2016 we created a testing swapbot in swapbot.tokenly.com site, but our bot is not showing in public TOKENLY API (we believe this happening because we don't pay the subscription) but this bot can be get from this URL  https://swapbot.tokenly.com/api/v1/public/bot/2115b238-70d4-4619-a9f5-41e2c82473c0. So, for now, we gonna hardcode the bot id and bot url for user mordorteam, our bot, we will remove this logic when the public API shows our bot.