Refactor: Fixed a null pointer exception when an artist doesn't have a TKY linked identity.
When the artist makes connection with a Fan, the identities TKY are shared, but, if the artist doesn't have a TKY linked identity, we got a null pointer exception when the linked process try to search a null value in TKY identity database. This is fixed, now we check if this value is null.
Also includes some clean code.