REFACTOR: fix a bug that shows a wrong image in ART communities menu.
The menu, in some cases shows a wrong image in the side menu, this image is taken from the selected identity, but, in some cases the side menu shows the default image. The problem was when the community is loaded, the module is not properly set and cannot find the selected identity.

Fix #6117